moved getMostRecentPastTerm function to separate file called PastTerm.js that also replaces the constants.js file
