--- frontend/src/constants.js
@@ -1,7 +1,0 @@
-// Term Planner Constants
-export const FEB = 1;
-export const JUN = 5;
-export const SEP = 8;
-export const MID_MONTH_START = 14;
-export const BEG_MONTH_START = 1;
-export const TERM_PAST_AMOUNT = 14;
