+++ frontend/src/pages/TermPlanner/ValidateTermPlanner.js
@@ -1,5 +1,5 @@
 import { plannerActions } from "../../actions/plannerActions";
-import { FEB, JUN, SEP, MID_MONTH_START, BEG_MONTH_START, TERM_PAST_AMOUNT } from "../../constants"
+import { getMostRecentPastTerm } from "./PastTerm"
 import axios from "axios";
 
 export const updateAllWarnings = (dispatch, plannerInfo, userInfo) => {
@@ -23,34 +23,6 @@ const validateTermPlanner = (payload) => {
         dispatch(plannerActions("TOGGLE_WARNINGS", data.courses_state));
       })
       .catch((err) => console.log(err));
-  };
-};
-
-// takes in startYear (int) and gets current date to return the most recent term that has past (week 2)
-// dictionary of year in degree and term most recent is returned 
-const getMostRecentPastTerm = (startYear) => {
-  let currTime = new Date();
-  let currYear = currTime.getFullYear();
-  let currMonth = currTime.getMonth();
-  let currDay = currTime.getDate();
-
-  const currDegreeYear = currYear - startYear + 1;
-
-  // session dates gathered from: https://www.student.unsw.edu.au/teaching-periods
-  let lastTermPast;
-  if ((currDay >= MID_MONTH_START + TERM_PAST_AMOUNT && currMonth == SEP) || currMonth > SEP) {
-    lastTermPast = 3;
-  } else if ((currDay >= BEG_MONTH_START + TERM_PAST_AMOUNT && currMonth == JUN) || currMonth > JUN) {
-    lastTermPast = 2;
-  } else if ((currDay >= MID_MONTH_START + TERM_PAST_AMOUNT && currMonth == FEB) || currMonth > FEB) {
-    lastTermPast = 1;
-  } else {
-    lastTermPast = 0;
-  }
-
-  return {
-    "Y": currDegreeYear,
-    "T": lastTermPast,
   };
 };
